Clean up TenorGif
