+++ build.gradle
@@ -74,5 +74,7 @@ dependencies {
     implementation 'org.mariadb.jdbc:mariadb-java-client:2.7.4'
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.0'
     testImplementation 'org.junit.jupiter:junit-jupiter-engine:5.8.0'
+    implementation 'com.fasterxml.jackson.core:jackson-databind:2.12.5'
+    implementation 'com.fasterxml.jackson.core:jackson-core:2.12.5'
 
 }
