+++ src/main/java/neptune/commands/CommandHandler.java
@@ -41,21 +41,10 @@ public class CommandHandler extends Helpers {
             commandRegistry.registerCommand(new commandBuilder().setCommand("MC").setCategory(CategoriesEnum.General).setRun(new MinecraftServerStatus()).setDescription("Check the state of a Minecraft server").build());
             commandRegistry.registerCommand(new commandBuilder().setCommand("ServerInfo").setCategory(CategoriesEnum.General).setRun(new GuildInfo()).setDescription("Some info about the current server").build());
             commandRegistry.registerCommand(new commandBuilder().setCommand("Log").setCategory(CategoriesEnum.Admin).setRequiredPermissions(new Permission[]{Permission.MANAGE_SERVER}).setRun(new Logging()).setDescription("Manage server logging").build());
-            commandRegistry.registerCommand(new commandBuilder().setCommand("Pat").setCategory(CategoriesEnum.Image).setRun(new Pat()).build());
-            commandRegistry.registerCommand(new commandBuilder().setCommand("Hug").setCategory(CategoriesEnum.Image).setRun(new Hug()).build());
-            commandRegistry.registerCommand(new commandBuilder().setCommand("Poke").setCategory(CategoriesEnum.Image).setRun(new Poke()).build());
-            commandRegistry.registerCommand(new commandBuilder().setCommand("Cuddle").setCategory(CategoriesEnum.Image).setRun(new Cuddle()).build());
-            commandRegistry.registerCommand(new commandBuilder().setCommand("Nom").setCategory(CategoriesEnum.Image).setRun(new Nom()).build());
-            commandRegistry.registerCommand(new commandBuilder().setCommand("Confused").setCategory(CategoriesEnum.Image).setRun(new Confused()).build());
             commandRegistry.registerCommand(new commandBuilder().setCommand("D10K").setCategory(CategoriesEnum.Fun).setDescription("10,000 Magical Effects").setRun(new D10K()).build());
             commandRegistry.registerCommand(new commandBuilder().setCommand("Powerlevel").setCategory(CategoriesEnum.Fun).setDescription("What does the scouter say about their power level?").setRun(new PowerLevel()).build());
-            commandRegistry.registerCommand(new commandBuilder().setCommand("Pout").setCategory(CategoriesEnum.Image).setRun(new Pout()).build());
             commandRegistry.registerCommand(new commandBuilder().setCommand("Senko").setCategory(CategoriesEnum.Image).setRun(new Senko()).build());
-            commandRegistry.registerCommand(new commandBuilder().setCommand("Stare").setCategory(CategoriesEnum.Image).setRun(new Stare()).build());
-            commandRegistry.registerCommand(new commandBuilder().setCommand("Neko").setCategory(CategoriesEnum.Image).setRun(new Neko()).build());
-            commandRegistry.registerCommand(new commandBuilder().setCommand("Shocked").setCategory(CategoriesEnum.Image).setRun(new Shocked()).build());
             commandRegistry.registerCommand(new commandBuilder().setCommand("Nya").setCategory(CategoriesEnum.Audio).setRun(new Nya()).setDescription("AUDIO: Nep goes Nya~").build());
-            commandRegistry.registerCommand(new commandBuilder().setCommand("Sleepy").setCategory(CategoriesEnum.Image).setRun(new Sleepy()).build());
             commandRegistry.registerCommand(new commandBuilder().setCommand("Breached").setName("Why the F*** was i breached").setCategory(CategoriesEnum.Fun).setRun(new WhyWasIBreached()).build());
             commandRegistry.registerCommand(new commandBuilder().setCommand("IsCaliforniaOnFire").setCategory(CategoriesEnum.Fun).setRun(new IsCaliforniaOnFire()).setDescription("Is California currently on fire?").build());
             commandRegistry.registerCommand(new commandBuilder().setCommand("Leaderboards").setCategory(CategoriesEnum.Utility).setRun(new Leaderboard()).setDescription("").build());
