+++ src/main/java/neptune/commands/ImageCommands/Tenor/Senko.java
@@ -1,15 +1,30 @@
 package neptune.commands.ImageCommands.Tenor;
 
 import neptune.commands.ICommand;
-import neptune.commands.TenorGif;
+import neptune.commands.ISlashCommand;
+import neptune.commands.ImageCommands.TenorGif;
 import net.dv8tion.jda.api.EmbedBuilder;
+import net.dv8tion.jda.api.MessageBuilder;
+import net.dv8tion.jda.api.entities.Message;
+import net.dv8tion.jda.api.events.interaction.SlashCommandEvent;
 import net.dv8tion.jda.api.events.message.guild.GuildMessageReceivedEvent;
+import net.dv8tion.jda.api.interactions.commands.build.CommandData;
 
-public class Senko extends TenorGif implements ICommand {
+public class Senko extends TenorGif implements ICommand, ISlashCommand {
     @Override
-    public void run(
-            GuildMessageReceivedEvent event, String messageContent) {
-        EmbedBuilder embedBuilder = getImageEmbed(event, "senko-san", false, null);
-        event.getChannel().sendMessage(embedBuilder.build()).queue();
+    public Message run(GuildMessageReceivedEvent event, String messageContent, MessageBuilder builder) {
+        EmbedBuilder embedBuilder = getImageEmbed("senko-san", false, null);
+        return builder.setEmbeds(embedBuilder.build()).build();
+    }
+
+    @Override
+    public CommandData RegisterCommand(CommandData commandData) {
+        return null;
+    }
+
+    @Override
+    public Message run(SlashCommandEvent event, MessageBuilder builder) {
+        EmbedBuilder embedBuilder = getImageEmbed("senko-san", false, null);
+        return builder.setEmbeds(embedBuilder.build()).build();
     }
 }
